+++ src/components/loader.js
@@ -14,7 +14,7 @@ const Loader = ({ onLoaderLoaded }) => {
 
     const { t: tr } = useTranslation();
     const appContext = useContext(AppContext);
-    const { setApiConfig, webConfig, setWebConfig, loadLanguages, setAllRoles, setAllPerms, loadMemberUIDs, loadDlogDetails } = useContext(AppContext);
+    const { setApiConfig, webConfig, setWebConfig, loadLanguages, setAllRoles, setAllPerms, setAllRanks, loadMemberUIDs, loadDlogDetails } = useContext(AppContext);
     const { themeSettings, setThemeSettings } = useContext(ThemeContext);
 
     const [isMember, setIsMember] = useState(false);
@@ -192,7 +192,7 @@ const Loader = ({ onLoaderLoaded }) => {
                 setAllPerms(memberPerms);
             }
             if (memberRanks) {
-                vars.ranks = memberRanks;
+                setAllRanks(memberRanks);
             }
 
             if (!useCache) {