+++ src/variables.js
@@ -33,9 +33,3 @@ var vtcLevel = 0; // 0: regular / 1: premium / 3: special
 exports.vtcLevel = vtcLevel;
 var userLevel = -1;
 exports.userLevel = userLevel;
-
-// DriversHub
-var perms = {};
-var ranks = {};
-exports.perms = perms;
-exports.ranks = ranks;